backend/requirements.txt
# ProeVira Backend - Dependencias Python

# Framework Web
flask>=2.3.0
flask-cors>=4.0.0

# Base de Datos
mysql-connector-python>=8.1.0

# Machine Learning y Análisis de Datos
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0

# Seguridad y Autenticación
bcrypt>=4.0.0

# Utilidades
python-dotenv>=1.0.0
werkzeug>=2.3.0
